requirements.txt
alembic==1.4.3
amqp==5.0.6
ansible==2.8.0
Babel==2.8.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
captcha==0.3
celery==5.0.5
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.2.1
dnspython==1.16.0
dominate==2.5.2
elasticsearch==7.9.1
eventlet==0.29.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==0.4.17
idna==2.10
importlib-metadata==3.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
numpy==1.19.4
pandas==1.1.4
paramiko==2.7.2
Pillow==8.3.1
prompt-toolkit==3.0.18
pycparser==2.20
pycryptodome==3.8.0
PyJWT==1.7.1
PyMySQL==0.10.0
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.4.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sphinxapi-py3==2.1.11
SQLAlchemy==1.3.18
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.25.10
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
zope.event==4.4
